Count revs for revno rather than use git-describe
We currently use 'git describe --always' to figure out the number of
commits there have been leading up to the HEAD commit.

However, if there are no tags in the repo, git-describe just returns
the sha hash of HEAD and we fail.

Use 'git rev-list HEAD | wc -l' instead so we're always #winning.

Change-Id: I345e0ee32189504276b3dfd3367057ce1d4a2b06