Fix "Needs to skip duplicated messages when using rabbitmq mirrored queue"
On using rabbitmq mirrored queue, openstack service may handle the same
message twice when rabbitmq is down after receiving a message,
dispatching to a specific method, and before sending ack to rabbitmq.
In this case, message is not deleted, then openstack service receives
same message after it reconnect to slave rabbitmq. This situation causes
an error since the message is already dispatched to a specific method.
This patch is to add unique_id to any message, let consumers store
unique_id of previous message and skip if openstack service
receive same message as previous one. One restriction is that this
bug fix is only dictionary type message.

Fixes bugs 1107064.

Change-Id: I5bfacadbdf7de8b34d6370b9aa869c271957692d