Only unpack replies on caller.
Opening the messages is unnecessary
in ZmqProxy, given that the msg_id
is already in the (zmq-specific)
message envelope. Additionally,
by not opening the messages, we
save a bit of complexity.

This avoids having to deserialize and
reserialize messages in ZmqProxy,
a necessary improvement for trusted-messaging
and should also have a markable performance
benefit.

Also cleans up msg_id handling a bit.

Change-Id: I31f365b9e1040503206b1d8320a0266a8121c821