Add support for directly stringifying VersionInfo.
glanceclient, and others, want to have a __version__ member in their
top-level __init__.py. This leads to attempting to process the
version information to realize the string at import time, rather
than at request time. If we make VersionInfo a thing which can
behave like a string, then, we should get the things we want without
the operational weirdness.

Change-Id: I3eb15006f50ed1155c85977aaae842677c5ec4bc