Cleanup docstring and print in install_venv_common
This commit cleans up usage in the docstring for:
run_command_with_code() to be more explicit in the description
of where the commands are being run.

It also updates the prints to be consistent in the usage of venv
vs virtualenv.

These changes were recommended in the review for change:
I68fe3449ed0fcc7fdfb84af2dcc77a39ebde7f01

Change-Id: I3cdb3d991503446d37c8fa6a3547585a4f8beb6a