Avoid using cfg in install_venv_common
Several people have complained that requiring oslo-config to be
installed in order to run install_venv seems like a chicken and egg
problem.

We're only using cfg for a boolean CLI option which is way overkill.
The truth is that cfg is really only interesting if you want to process
options from both the command line and configuration files. Clearly
that's not the case here.

Just use argparse instead of cfg.

Change-Id: Ib602686b554c3548f4be3d4a69d128c5fef7e8cf