Revert "Fast serialization of RPC envelopes for Kombu"
This reverts commit 22c497097b0d0bd461be40a7a03290aa0b4179f2.

I'm not convinced that this isn't just a micro-optimization when
compared to the rest of the work that a given OpenStack service performs
as the result of these messages.

The implementation is also problematic.  It depends on using a '\0' byte
as a separator and assumes that '\0' will not exist anywhere else in the
message.  This seems to be asking for trouble.  I'm not convinced that
this can *never* happen and if it did, this code would break.  Further,
if a user could get a '\0' in a message directly with user-supplied
input, this could result in a security vulnerability.

Change-Id: Ib3163ca98f568bf9f789d4b64bcc6d72e0fcb459