Improves Logging for for rpc method failures
Provides a decorator for rpc calls.  The decorator catches exceptions,
and logs it with a traceback message, along with the rpc-topic and
rpc-method being called.

This logging is on by default but can be disabled by the caller if
required by passing: '_log_exc=False' as a keyword argument when calling
the decorated method that has been decorated.

Works towards: bug #1148516

Change-Id: I29a3b1b97c6114c4479e2b71c1257c4d72131535