Add 'is' operator to extra specs ops.
Boolean values for capabilities don't work because extra_specs are
all converted to unicode. The scheduler will then check, for example,
if the boolean 'True' is equal to the unicode string 'True', and will
always return False. This patch allows admins to specify '<is> True'
in extra_specs, which will compare successfully to boolean True.

Fixes bug: 1146306

Change-Id: Id0e6dcfb71eb0943a16bba551ec23c4d57206550