Update install_venv_common to use patch -N.
This patch makes it possible to call the post_process function
in install_venv_common (which currently only patches files)
multiple times.

The motivation here is that I would like to use this to patch
virtual env's created by tox. Tox doesn't provide a good hook
to call post_process (which patches files) only once so the
idea here is we'll just let it call post_process all the time...
and for distro's where patching is used to fix the venv the
patch -N option will make it skip patches on subsequent runs.

Change-Id: Ib0a36e49dd1884de131c569fe8bb66fe20285e97