Don't check exit codes when patching.
Similar to what we did in 427769c we need to ignore checking
exit codes for patch -N command so that they can be run multiple
times.

The motivation here is that I would like to use this to patch
virtual env's created by tox. Tox doesn't provide a good hook
to call post_process (which patches files) only once so the
idea here is we'll just let it call post_process all the time...

Change-Id: I866e7a3e1fd492d7d4c08eb55cde02f2dfa16c9a