Improves Logging for for rpc method timeouts
This patch moves the traceroute for an rpc timeout from inside an iterator, which gave a useless trace back, into the main flow of the program.

It also adds the rpc method being called and the topic used to the exception's message.

When the caller logs the message higher up the stack, the log information will be more useful.

Finally it removes the timeout logging in the amqp.py module, in the
spirit of bug #1137994 and https://review.openstack.org/#/c/23295/

Works towards: bug #1148516

Change-Id: I29a3b1b97c6114c4479e2b71c1257c4d72131535