Fix log.py to check the existence of log conf file
The log conf file could be nonexistent or invalidate, but
the code throws out 'NoSectionError' in either case, which
is not accurate.

Added a check for the existence of log conf file. If not exists,
throw out a RuntimeError. If the file is missing specific
sections and/or options, the code throws out corresponding
exceptions as before.

Fixes: bug #1154245
Change-Id: I8bc8f5392263041168bb57d80ec8ecd577a4f47c