Fix log.py to check the access of log conf file
The log conf file could be nonexistent or invalid, but
the code throws out 'NoSectionError' in either case, which
is not accurate.

Added a check for the existence and access of log conf file,
and put the file check and parsing into a try-except block.
The RuntimeError, if any, will contain any error.

Fixes: bug #1154245
Change-Id: I8bc8f5392263041168bb57d80ec8ecd577a4f47c