Eliminate multiple contexts in zmq call()
Because we need compatibility with Grizzly (N-1 support), this should be
reduced to two copies for Havana, before stripping down to a singular
copy for I.

The '-reply' request has its context key removed, with the
msg option maintaining one extra copy of context until I.
A TODO has been entered to clean this up following the Havana release.

This fix eliminates an extra JSON double-deserialization routine.
The double-serialization of context is still performed for Grizzly
compatibility.

fixes bug 1156956

Change-Id: I071e382ec9c7e0df3794fe9ca574a5722f8910a6