clarify --log-file comments
Clarify that --log-dir is only applicable for
relative --log-file paths.

Also change the description of --log-file, to be less specific.
By changing "If not set" to "If no default is set", it
allows for a better description given the cumulative nature of
these config settings.  Distributors can use this for example
to set a default in a distribution wide config file that is
then indicated by a commented default in a user editable file.
I.E. the following config file is now more accurate:

 # Name of log file to output to.
 # If no default is set, logging will go to stdout.
 # log_file = /some/dist/default/set/elsewhere.conf

Change-Id: I9a39511b6d6cfff907471ab7a6f31c34592a1015