lockutils: improve the external locks test
Use a separate directory for the lockutils external locks and for the
test's flocks which are used to check that serialization is actually
occurring. By using the same directory, we can't test the code in
lockutils which auto-creates this directory because the dir gets
deleted by a process by lockutils cleanup while another directory
is using it for flocks.

Also, assume the the handles directory has been created in the parent
rather than having each child attempt to create it.

Related, add a try/finally block so that when a child process throws an
exception it immediately exits rather than deleting the temporary
directories created by the parent.

Change-Id: I32d7e8e05fb3f22cf38fa586f8bc97646c83f182