Fix test_write_git_changelog
The test changes current directory to a temporary one
and ChangeLog file is created here. That's why the file
doesn't reside in project's root directory.

The correct behavior is to look for ChangeLog at cwd as is
done in test_generate_authors for AUTHORS file.

Use openstack.__file__ do determine the root directory
of openstack package.

Change-Id: Ic7c9b01485a08b8921c88ac624e0aecd9409fa7c
Fixes: bug #1161362