Add rpc method namespace support.
RPC endpoints already had the ability to expose multiple APIs ... sort
of.  You could pass multiple callback objects to the dispatcher and it
would check all of them for a method call.

This patch adds the ability to set a namespace on a callback object.
This makes exposing multiple APIs a bit more like you would expect it to
work.  You can invoke a method on a specific callback object, as opposed
to having it check all of them for the method.

This will allow you to create, manage, and version APIs without any
potential conflicts with other APIs being exposed by the same endpoint.
An example of where I would like to use this is in Nova, where we have
some methods that we would like to expose on *all* rpc endpoints.

This includes no public API changes and is fully backwards compatible.

Implement blueprint rpc-multi-api.

Change-Id: Ief4433e2e1c32cfb05b4cd27b87fe32b40f4341d