Allow overriding the formatter used for logging
Adds a new config option 'logging_formatter' which allows overriding the
formatter added to all logging handlers.  By default a DefaultFormatter
will be used which pulls its format from CONF.log_format.  This avoids
the current behaviour of needing to explicitly set log_format='' in
order to use the LegacyFormatter.  And it opens the way for projects
using this module to specify their own formatters.

Bug 1167388

Change-Id: I281ed7fe3448403f27a4cd318fc21915027906cc