Allow overriding the formatter used for logging
This patch attempts to reconcile two currently incompatible ways of
configuring log formatting.  Because log_format had a default set it was
always used in favor of the context aware formatter.  In order to
deprecate the option it now defaults to None so it is overridable, but
it should be removed at some point.

Adds a new config option 'logging_formatter' which allows overriding the
formatter added to all logging handlers.  By default a ContextFormatter
will be used which pulls its format from current logging options.  This
method of choosing a formatter should allow each project to customize
it to suit their needs with minimal duplication.

Bug 1167388

Change-Id: I281ed7fe3448403f27a4cd318fc21915027906cc