Remove rootwrap from IPC directory creation
Removes processutils/rootwrap dependency.
Fixes bug 1160420 and partially fixes 1180631.

The attempt will still be made to create
/var/run/openstack, but will give a reasonable
error when it happens.

If the directory exists, but permissions are denied,
we should also give a reasonable error.

A patch will be submitted to devstack
that will set and create the IPC directory as necessary.
Packagers should do this as well.

Change-Id: I7814320adbfd0b2fe54cc2e523452c18ab2e7687