RPC: Add support for optional message signing
Wires in SecureMessage so that signing can be optionally used.
Signing messages is always attempted but failing to create a signed message is
currently not fatal. In case of failure the message is sent unsigned.

Once all services has been converted to support signing, we can start switching
on mandatory signing.

Currently the service name needs to be a global due to the way services are
constructed.

Change-Id: Ie565bbaf61a69a4e526e78e723efecb5e2712877