Added convenience APIs for lockutils
The lock_file_prefix for each project doesn't need to be configurable or
frequently changed. This patch provides a convenience API which returns
a partial object of the synchronized decorator to avoid passing the prefix
each time locks are used.

The set_defaults method is also provided to change the default value of
lock_path when needed.

Fixes bug #1065524

Change-Id: I7b67f0a482da4be6d53a70db5bbd22dc91bdc10c