Add serialization hooks to RpcProxy
This adds hooks in RpcProxy to serialize arguments on the way out
and to deserialize a response on the way back in from a call. The
base serialization implementation can call a to_primitive() method
on the argument, if it has one, making it likely to work for
actual users. The deserialization is just a pass-through hook
right now, which can be implemented appropriately in a subclass.

Change-Id: I0a33baddee3e28dfc47100eb3216b679558b0bdd