Add serialization hooks to RpcDispatcher
This adds hooks for deserializing arguments before calling a
dispatched method, and for serializing the result before it is
passed back over RPC to the caller. The base serialization calls
a to_primitive() method on the result, if it exists. The
deserialization is just a pass-through hook right now, which can
be implemented appropriately in a subclass.

Related to blueprint rpc-object-serialization

Change-Id: I845727552097893d74f2ee7f5bb63b15460acdf7