Add 't', 'y', and `strict` to `bool_from_string`
't' and 'y' are commonly used for booleans as well, and since it's used in
Nova already, probably makes sense to add common support for it.

`strict` allows callers to detect anonmolous values so they can be handled
separately. This is particularly useful when parsing input from API calls
where bad values might result in 4XX errors instead of being treated as False.

Change-Id: I6218a1895f72cfbfd81776602e63ca3d093475e0