Handle exec_dirs for argument in Killfilter
KillFilter currently expects an absolute path
to be specified for the process name to kill. This
is inconvenient when the exact location of the running
binary is not known or differs accross installs.

Extend KillFilter to accept also commands to be
killed that are in exec_dirs and that match the
(now non-absolute) path given as KillFilter argument.

Change-Id: I6b90206b587ff3f949af2c256a78ca21af31867a