Use zero-copy in ZeroMQ proxy (zmq-rpc-receiver)
Now, messages will not be copied into the Python memory space
from the C library. The 'topic' key *will* be copied, to make
use of several string/bytes functions that are available to us.
(Potentially, we might be able to work with this as a memoryview
in the future as a further optimization?)

Implements blueprint zmq-zero-copy

Change-Id: I20a700867ac642b1a1788b31e3a21f8a299c97c8