Move wsgi module to deprecated package
As we are planning to deprecate the wsgi module, move it
out of the main openstack/common folder to the new
"deprecated" folder.

Move some of the middleware classes into the middleware
package, since those are still useful and should be
maintained.

There's still work to do to remove the dependency of the
size limit middleware on the Request class in the now
deprecated wsgi module.

Refer to the mailing list thread at
http://lists.openstack.org/pipermail/openstack-dev/2013-May/009666.html
for additional details.

Change-Id: I05ff583ef37174835393905a3066390f400ffdcf
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>