Add a monkey-patching util for sqlalchemy-migrate
Nova and other projects use sqlalchemy-migrate for DB schema
migrations. Unfortunately, this project looks like to be dead,
but have some important bugs which makes lives of OpenStack
developers harder (e. g. creation of a new unique constraint
in SQLite leads to deletion of all existing unique constraints).

Nova has some workarounds for bugs and limitations of sqlalchemy-migrate,
though it would be nice to have those directly in sqlalchemy-migrate
(at least in form of a monkey-patch for now). Oslo seems to be
a good place to store this monkey-patch, so Nova and other projects
could reuse it.

This patch:
  - makes it possible to use the unified drop_unique_constraint()
    function for SQLite backend
  - fixes a bug in sqlalchemy-migrate that leads to deletion of
    existing unique constraints of a table when a new one is added
    (SQLite backend)

Here is the corresponding 'proof-of-concept' patch in Nova:
https://review.openstack.org/#/c/34526/2

Change-Id: Ifac07abac3814b3ea4dea5840b17a711f4b24b8d