Don't raise Timeout on no-matchmaker results
We were raising a Timeout when no matchmaker results
were returned. While arguably legitimate (especially with
'call'), there are places in the code where this may
not be caught (or cared about) on 'cast'.

When sending a cast, we'll simply fail to deliver the message
and log a warning. Using call() will still raise a Timeout.

This bug would generally surface when there were no consumers
for a topic (in the view of the selected matchmaker)

Fixes bug 1186310

Change-Id: I3c351a4ca921aa804caeaf0751351e036268a3bb