Default to sending messages with RPC envelope
The ZMQ driver was, per Grizzly, defaulting to
sending messages with the RPC envelope disabled.

When the RPC envelope was turned on globally,
this did not take immediate effect for the ZMQ
driver due to how optional-envelope support
was implemented for notifications.

This patch simply changes the envelope kwarg default
from False to True. The notify() in rpc/common.py
defaults its envelope parameter to False, overriding
this default when notifications are sent.

Fixes bug 1187036

Change-Id: I53e14c6937b0fc3879dcb8fbed19310c244f2e61