Allow exceptions to hop up cells
When an exception happens in an RPC call using nova cells, it can travel
back up several RPC boundaries. eg, child-cell=>parent-cell=>nova-api

Before this patch if an exception (eg. AggregateNotFound) was raised in
the bottom layer, the next layer up would turn it into a special
exception ( AggregateNotFound_Remote ), then in the final layer, it would
see this as an unrecognizable exception and raise RemoteException.

After this patch, at each layer where the expeption is deserialized,
it'll recognie exceptions with the _Remote postfix, and leave them as
they are, instead of turning them into RemoteExceptions.

It also preserves the exception's original __module__ now.

Change-Id: I158a80f1cec20d3e1805b565ffddaffd7a15295b