Make lock_file_prefix optional
Currently, the lock_file_prefix says it should end with an hypen if
specified. It seems that it was intended to be optional. Since that
behavior makes sense, this patch changes the synchronized's signature
and makes lock_file_prefix optional.

With this patch, the prefix will not be required to end with an hypen
anymore, such sign will be added to the prefix if specified and if not
already present.

This patch doesn't change synchronized's behavior at all, which means it
doesn't break backward compatibility.

Change-Id: I8841da1e8d105aac4a86b247228a9c935e00afc5