Add a default entry of sudo to execute root_helper.
The current execute function in processutils:execute doesn't
set a default root_helper.  Since we're only setting this when
run_as_root is True, it seems like it would be ok to add a default
of the currently accepted root_helper.

This won't impact folks that are currently using the code with
a different root_helper specified as they need to set this currently
anyway.

Change-Id: Iead33772326e63a26d533bd256b8caa30d6451bc