Remove unnecessary timeout logging.
rpc.call timeouts are not necessarily fatal, so don't log an exception
with a traceback for every single one.  Let the caller decide what to
do.

This is a partial backport of 229b815fb72dccf998b18a0ad019cc61ccb888d8
from master, which removed this along with a bunch of other changes.

I noticed this during a sync to nova, where this line was getting added,
so it must have been removed directly in nova at some point.  It really
shouldn't be there, so let's just remove it from the original code.

Change-Id: Ide4f26865d8f2ce0630cc8824fbea6f17aa262fa