Improve usability when backdoor_port is nonzero
Users who may not know that configuring a backdoor_port with 0
allows multiple services to be enabled for the eventlet backdoor
or who simply want a more predictable port assignment might like
this patch. If the specified port is in use, it is incremented
until a free port is found. This is a backdoor_port collision
recovery scheme as opposed to the collision failure scheme
that exists today.

This related to I95fdb5ca: Add support for backdoor_port to be
returned with a rpc call.

Change-Id: I7ec346db3575995fa15483b617eea34c1e003bb0