Better socket termination for ZeroMQ
We were telling ZeroMQ to close sockets,
but only after successfully sending them.
If a socket timed out, we'd drop the reference
from Python, but the ZeroMQ library was still
keeping these sockets alive in the background.

Instead of using eventlet's Timeout,
we can simply use the linger argument to
terminate sockets if they're alive for too
long. If the message sends before the timeout,
all is good. Otherwise, it is dropped.

Change-Id: I07c7996ee76873ee762582e58dbbec05bc7de66c