Avoid shadowing Exception 'message' attribute
The Exception class has a 'message' attribute:

  >>> Exception('foo').message
  'foo'

however, we follow a pattern that results in the attribute being
shadowed by a class attribute:

  >>> class MyException(Exception):
  ...     message = 'Foo %s'
  ...     def __init__(self, bar):
  ...         message = self.message % bar
  ...         super(MyException, self).__init__(message)
  ...
  >>> MyException('bar').message
  'Foo %s'

whereas, we obviously want behaviour like this:

  >>> class MyException(Exception):
  ...     msg_fmt = 'Foo %s'
  ...     def __init__(self, bar):
  ...         message = self.msg_fmt % bar
  ...         super(MyException, self).__init__(message)
  ...
  >>> MyException('bar').message
  'Foo bar'

Avoid this shadowing by using 'msg_fmt' as the name of the class
attribute. Add a test which fails without this fix.

Change-Id: I304f31edec20b1293cf6c8de931415590fde6752