Introduces Basic Report and Model
Introduces the classes for basic reports and the base data model
class to hold data for the report.  A basic report is composed of
sections, each of which has a generator (which produces the model)
and a view (which converts the model into a given representation).
Models may have an associated view, as well, to allow for sub-models
to have specific views attached to them.  This is the first part
of the Guru Meditation Report Framework.

Change-Id: Ic3ad1fd595da8742b5e7d4349a23132fc560f292
Implements: bp guru-meditation-report